# Conflicts:
#	GoogleMapsComponents/Maps/Map.cs

+++ GoogleMapsComponents/Maps/Map.cs
@@ -159,9 +159,9 @@ public class Map : EventEntityBase, IJsObjectRef, IAsyncDisposable
     /// The heading value is measured in degrees (clockwise) from cardinal direction North.
     /// </summary>
     /// <returns></returns>
-    public Task<int> GetHeading()
+    public Task<float> GetHeading()
     {
-        return _jsObjectRef.InvokeAsync<int>("getHeading");
+        return _jsObjectRef.InvokeAsync<float>("getHeading");
     }
 
     /// <summary>
@@ -169,7 +169,7 @@ public class Map : EventEntityBase, IJsObjectRef, IAsyncDisposable
     /// </summary>
     /// <param name="heading"></param>
     /// <returns></returns>
-    public Task SetHeading(int heading)
+    public Task SetHeading(float heading)
     {
         return _jsObjectRef.InvokeAsync("setHeading", heading);
     }
